requirements.txt
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.1.0
xlrd>=2.0.1
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.1.0
pyyaml>=6.0
xlsxwriter>=3.0.0
flask>=2.3.0
flask-cors>=4.0.0
yfinance>=0.2.0
pmdarima>=2.0.0
statsmodels>=0.13.0
hmmlearn>=0.3.0
forex-python>=1.8.0
click>=8.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.28.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
flask-login
python-dotenv>=1.0.0
Flask-Babel>=3.0.0
Babel>=2.12.0
